application/requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==2.16.1
attrs==22.2.0
azure-identity==1.12.0
cffi==1.15.1
cryptography==39.0.2
decorator==5.1.1
Flask==2.2.3
gunicorn==20.1.0
Jinja2==3.1.2
numpy==1.24.2
openai==0.27.0
openapi-schema-pydantic==1.2.4
pydantic==1.10.7
pyodbc==4.0.35
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
retry==0.9.2
rsa==4.9
SQLAlchemy==1.4.46
tenacity==8.2.2
tiktoken==0.3.3
urllib3==1.26.14
Werkzeug==2.2.3
